shachotwo/z_その他/archive/shachotwo-マーケAI/requirements.txt
# Web framework
fastapi==0.115.*
uvicorn[standard]==0.34.*
jinja2==3.1.*
python-multipart==0.0.*

# Google APIs (Sheets, Gmail, Calendar)
google-api-python-client==2.160.*
google-auth-httplib2==0.2.*
google-auth-oauthlib==1.2.*

# Browser automation (form submission)
playwright==1.49.*

# Scraping
beautifulsoup4==4.12.*
httpx==0.28.*
lxml==5.3.*

# PDF generation
weasyprint==63.*

# LLM
google-generativeai==0.8.*

# Database
supabase==2.11.*

# Data validation
pydantic==2.10.*
pydantic-settings==2.7.*

# Environment
python-dotenv==1.0.*
